Move function to encode an image to converters
