+++ jupyter_docx_bundler/converters.py
@@ -1,9 +1,31 @@
+import base64
 import os
 import re
 import tempfile
 
 from nbconvert import HTMLExporter, preprocessors
 import pypandoc
+
+
+def encode_image_base64(filepath):
+    """Encode an image as a base64 string
+
+    Parameters
+    ----------
+    filepath : str
+        Filepath of the image file
+
+    Returns
+    -------
+    dict
+        Dictionary with identifier as key and base64-encoded data as value.
+
+    """
+    key = 'image/' + os.path.splitext(filepath)[1][1:]
+    with open(filepath, 'rb') as image:
+        data = base64.b64encode(image.read()).decode('utf8')
+
+    return {key: data}
 
 
 def preprocess(content):
